+++ bastion/provider.go
@@ -60,6 +60,7 @@ func Provider() *schema.Provider {
 			"wallix-bastion_application_localdomain":               resourceApplicationLocalDomain(),
 			"wallix-bastion_application_localdomain_account":       resourceApplicationLocalDomainAccount(),
 			"wallix-bastion_authdomain_ldap":                       resourceAuthDomainLdap(),
+			"wallix-bastion_authdomain_mapping":                    resourceAuthDomainMapping(),
 			"wallix-bastion_authorization":                         resourceAuthorization(),
 			"wallix-bastion_checkout_policy":                       resourceCheckoutPolicy(),
 			"wallix-bastion_cluster":                               resourceCluster(),