+++ bastion/resource_ldapmapping.go
@@ -8,6 +8,7 @@ import (
 	"net/url"
 	"strings"
 
+	govers "github.com/hashicorp/go-version"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	bchk "github.com/jeremmfr/go-utils/basiccheck"
@@ -47,8 +48,17 @@ func resourceLdapMapping() *schema.Resource {
 	}
 }
 func resourceLdapMappingVersionCheck(version string) error {
-	if bchk.InSlice(version, defaultVersionsValid()) {
+	if bchk.InSlice(version, []string{VersionWallixAPI33, VersionWallixAPI36}) {
 		return nil
+	}
+	if vers, err := govers.NewVersion(version); err == nil {
+		versionResourceRename, _ := govers.NewVersion(VersionWallixAPI38)
+		if vers.GreaterThanOrEqual(versionResourceRename) {
+			return fmt.Errorf(
+				"resource wallix-bastion_ldapmapping not available with api version %s\n"+
+					" use wallix-bastion_authdomain_mapping instead",
+				version)
+		}
 	}
 
 	return fmt.Errorf("resource wallix-bastion_ldapmapping not available with api version %s", version)